requirements.txt
# Core AI and ML
openai>=1.0.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=11.0.0
pydantic>=2.5.0

# Database and Storage
sqlite3
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# HTTP Requests and APIs
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# Task Queue and Background Jobs
celery>=5.3.0
redis>=5.0.0

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
pre-commit>=3.5.0

# Utilities
click>=8.1.0
rich>=13.7.0
schedule>=1.2.0
pytz>=2023.3

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0 